remove the claim page, now we claim from allocation page
